requirements.txt
langchain-google-genai
rich
python-dotenv